streamlit-app/requirements.txt
streamlit==1.29.0
pandas==2.1.0
numpy==1.26.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
matplotlib==3.7.2
opencv-python==4.8.0.76
Pillow==10.0.0
pyyaml==6.0
huggingface-hub==0.21.3
